+++ CHANGES.rst
@@ -18,6 +18,7 @@ Bugfixes
 - Do not show borders above/below the message in registration emails
   unless both the header and body blocks are used (:issue:`3151`)
 - Roll-back the database transaction when an error occurs.
+- Fix rendering of the LaTeX error box (:issue:`3163`)
 
 
 Version 2.0rc1